refactor: improve component organization and code quality
* refactor: organize components into feature-based directories

- Remove unused HelloWorld.vue component
- Create feature-based directory structure:
  - task/ for task-related components
  - dialog/ for dialog components
  - form/ for form components
- Update import paths in all affected components:
  - App.vue
  - TaskList.vue
  - TaskItem.vue

* refactor(TaskItem): simplify dialog opening methods

- Replace inline dialog opening logic with dedicated methods for better readability
- Update button click handlers to use new methods for opening edit and delete dialogs
